+++ pymongo/common.py
@@ -113,6 +113,9 @@ UNAUTHORIZED_CODES = (13, 16547, 16548)
 # From the driver sessions spec.
 _MAX_END_SESSIONS = 10000
 
+# Default value for srvServiceName
+SRV_SERVICE_NAME = "mongodb"
+
 
 def partition_node(node):
     """Split a host:port string into (host, int(port)) pair."""
@@ -626,6 +629,7 @@ URI_OPTIONS_VALIDATOR_MAP = {
     'w': validate_non_negative_int_or_basestring,
     'wtimeoutms': validate_non_negative_integer,
     'zlibcompressionlevel': validate_zlib_compression_level,
+    'srvservicename': validate_string
 }
 
 # Dictionary where keys are the names of URI options specific to pymongo,